Remove deprecated register
In c++11, register has been deprecated, and
all of skia is now built as gnu++11

From the documentation:
A register specifier is a hint to the implementation
that the variable so declared will be heavily used.
[ Note: The hint can be ignored and in most implementations
it will be ignored if the address of the variable is taken.
This use is deprecated (see D.2) ]

Change-Id: Ic7432b034037fea273c4a9c09f0ca4a728f7eb88
Signed-off-by: DHOMD <DHO@dho.im>